feat: move contact and quote logic to backend (ATA-7)
Ports the Mailgun email sending, rate limiting, and quote estimate
calculation from the frontend's API routes into real backend routes
behind verifyApiKey. The frontend's /api/contact and /api/quote become
thin proxies that only validate CSRF and forward the request (including
the client IP via X-Forwarded-For, so backend rate limiting still works
per-IP).
