+++ apps/backend/package.json
@@ -7,12 +7,16 @@
     "dev": "tsx watch src/index.ts",
     "build": "tsc",
     "start": "node dist/index.js",
-    "seed:post": "tsx scripts/seed-post.ts"
+    "seed:post": "tsx scripts/seed-post.ts",
+    "test": "node --import tsx --test src/**/*.test.ts"
   },
   "dependencies": {
     "@modelcontextprotocol/sdk": "^1.30.0",
     "dotenv": "^16.6.1",
     "express": "^4.21.0",
+    "express-rate-limit": "^7.4.0",
+    "form-data": "^4.0.4",
+    "mailgun.js": "^10.4.0",
     "zod": "^4.4.3"
   },
   "devDependencies": {
