+++ apps/backend/src/__tests__/quoteCalculator.test.ts
@@ -1,4 +1,6 @@
-import { calculateQuoteEstimate, QuoteInput } from '../utils/quoteCalculator';
+import { describe, it } from 'node:test';
+import assert from 'node:assert/strict';
+import { calculateQuoteEstimate, type QuoteInput } from '../services/quoteCalculator.js';
 
 const baseInput: QuoteInput = {
   budget: 5_000_000,
@@ -9,49 +11,49 @@ const baseInput: QuoteInput = {
 describe('calculateQuoteEstimate', () => {
   it('returns a valid estimate with default locale (es)', () => {
     const result = calculateQuoteEstimate(baseInput);
-    expect(result.requestedBudget).toBeGreaterThanOrEqual(1_000_000);
-    expect(result.recommendedBudget.min).toBeGreaterThan(0);
-    expect(result.recommendedBudget.max).toBeGreaterThan(result.recommendedBudget.min);
-    expect(result.phases).toHaveLength(5);
-    expect(result.suggestions.length).toBeGreaterThan(0);
+    assert.ok(result.requestedBudget >= 1_000_000);
+    assert.ok(result.recommendedBudget.min > 0);
+    assert.ok(result.recommendedBudget.max > result.recommendedBudget.min);
+    assert.strictEqual(result.phases.length, 5);
+    assert.ok(result.suggestions.length > 0);
   });
 
   it('returns localized labels for es locale', () => {
     const result = calculateQuoteEstimate(baseInput, 'es');
-    expect(['Baja', 'Media', 'Alta']).toContain(result.complexity);
-    expect(['Baja', 'Media', 'Alta', 'Muy alta']).toContain(result.timelinePressure);
+    assert.ok(['Baja', 'Media', 'Alta'].includes(result.complexity));
+    assert.ok(['Baja', 'Media', 'Alta', 'Muy alta'].includes(result.timelinePressure));
   });
 
   it('returns localized labels for en locale', () => {
     const result = calculateQuoteEstimate(baseInput, 'en');
-    expect(['Low', 'Medium', 'High']).toContain(result.complexity);
-    expect(['Low', 'Medium', 'High', 'Very high']).toContain(result.timelinePressure);
+    assert.ok(['Low', 'Medium', 'High'].includes(result.complexity));
+    assert.ok(['Low', 'Medium', 'High', 'Very high'].includes(result.timelinePressure));
   });
 
   it('enforces minimum budget', () => {
     const result = calculateQuoteEstimate({ ...baseInput, budget: 100 });
-    expect(result.requestedBudget).toBe(1_000_000);
+    assert.strictEqual(result.requestedBudget, 1_000_000);
   });
 
   it('clamps timeline to valid range', () => {
     const short = calculateQuoteEstimate({ ...baseInput, timelineMonths: 0 });
-    expect(short.requestedTimelineMonths).toBe(1);
+    assert.strictEqual(short.requestedTimelineMonths, 1);
 
     const long = calculateQuoteEstimate({ ...baseInput, timelineMonths: 100 });
-    expect(long.requestedTimelineMonths).toBe(24);
+    assert.strictEqual(long.requestedTimelineMonths, 24);
   });
 
   it('returns phases that sum to 100%', () => {
     const result = calculateQuoteEstimate(baseInput);
     const totalPercentage = result.phases.reduce((sum, p) => sum + p.percentage, 0);
-    expect(totalPercentage).toBeCloseTo(1.0, 5);
+    assert.ok(Math.abs(totalPercentage - 1.0) < 1e-9);
   });
 
   it('handles all technology types', () => {
     const technologies = ['web_app', 'integrations', 'mobile_app', 'other'] as const;
     for (const tech of technologies) {
       const result = calculateQuoteEstimate({ ...baseInput, technology: tech });
-      expect(result.recommendedBudget.min).toBeGreaterThan(0);
+      assert.ok(result.recommendedBudget.min > 0);
     }
   });
 
@@ -59,17 +61,17 @@ describe('calculateQuoteEstimate', () => {
     const short = calculateQuoteEstimate({ ...baseInput, timelineMonths: 1 }, 'en');
     const long = calculateQuoteEstimate({ ...baseInput, timelineMonths: 12 }, 'en');
     const pressureOrder = ['Low', 'Medium', 'High', 'Very high'];
-    expect(pressureOrder.indexOf(short.timelinePressure)).toBeGreaterThanOrEqual(
-      pressureOrder.indexOf(long.timelinePressure)
+    assert.ok(
+      pressureOrder.indexOf(short.timelinePressure) >= pressureOrder.indexOf(long.timelinePressure)
     );
   });
 
   it('phase amounts are rounded to nearest thousand', () => {
     const result = calculateQuoteEstimate(baseInput);
     for (const phase of result.phases) {
-      expect(phase.minAmount % 1000).toBe(0);
-      expect(phase.maxAmount % 1000).toBe(0);
-      expect(phase.averageAmount % 1000).toBe(0);
+      assert.strictEqual(phase.minAmount % 1000, 0);
+      assert.strictEqual(phase.maxAmount % 1000, 0);
+      assert.strictEqual(phase.averageAmount % 1000, 0);
     }
   });
 });
