+++ apps/backend/src/types/locale.ts
@@ -1,0 +1,6 @@
+export type Locale = 'es' | 'en';
+
+export const DEFAULT_LOCALE: Locale = 'es';
+
+export const isSupportedLocale = (value: unknown): value is Locale =>
+  value === 'es' || value === 'en';
