+++ apps/backend/src/utils/sanitize.ts
@@ -1,0 +1,12 @@
+const HTML_ESCAPE_MAP: Record<string, string> = {
+  '&': '&amp;',
+  '<': '&lt;',
+  '>': '&gt;',
+  '"': '&quot;',
+  "'": '&#x27;',
+};
+
+const HTML_ESCAPE_REGEX = /[&<>"']/g;
+
+export const escapeHtml = (str: string): string =>
+  str.replace(HTML_ESCAPE_REGEX, (char) => HTML_ESCAPE_MAP[char] || char);
