+++ apps/frontend/package.json
@@ -21,10 +21,7 @@
     "dompurify": "^3.1.7",
     "dotenv": "^16.6.1",
     "express": "^4.21.0",
-    "express-rate-limit": "^7.4.0",
-    "form-data": "^4.0.4",
     "framer-motion": "^11.9.0",
-    "mailgun.js": "^10.4.0",
     "marked": "^14.1.2",
     "next": "14.2.35",
     "nodemailer": "^6.9.15",
