+++ apps/frontend/src/server/backendClient.ts
@@ -9,6 +9,18 @@ async function backendFetch(path: string): Promise<Response> {
   });
 }
 
+async function backendPost(path: string, body: unknown, clientIp?: string): Promise<Response> {
+  return fetch(`${BACKEND_URL}${path}`, {
+    method: 'POST',
+    headers: {
+      Authorization: `Bearer ${BACKEND_API_KEY}`,
+      'Content-Type': 'application/json',
+      ...(clientIp ? { 'X-Forwarded-For': clientIp } : {}),
+    },
+    body: JSON.stringify(body),
+  });
+}
+
 export async function listPosts(): Promise<PostData[]> {
   const res = await backendFetch('/posts');
   if (!res.ok) {
@@ -26,4 +38,15 @@ export async function getPostBySlug(slug: string): Promise<PostData | null> {
     throw new Error(`Backend GET /posts/${slug} failed: ${res.status}`);
   }
   return res.json();
+}
+
+export async function sendContactMessage(
+  payload: { name: string; email: string; message: string },
+  clientIp?: string
+): Promise<Response> {
+  return backendPost('/contact', payload, clientIp);
+}
+
+export async function submitQuote(payload: Record<string, unknown>, clientIp?: string): Promise<Response> {
+  return backendPost('/quote', payload, clientIp);
 }
